pyproject.toml
[tool.poetry]
name = "agent-reflexion"
version = "0.1.0"
description = "Reflective agent capable of self-critique and improvement."
authors = ["princeGedeon <guedjegedeon03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^5.13.2"
langchain = "^0.3.7"
langgraph = "^0.2.46"
langchain-openai = "^0.2.8"
grandalf = "^0.8"
langsmith = "^0.1.143"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
